requirements.txt
anyio==3.6.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bagpy==0.5
bitstring==4.0.1
catkin-pkg==0.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
distro==1.8.0
docutils==0.19
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.95.0
ffmpeg-python==0.2.0
fonttools==4.39.3
future==0.18.3
gnupg==2.3.1
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.0.3
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.1
Pillow==9.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py3rosmsgs==1.18.2
pycryptodomex==3.17
pydantic==1.10.7
Pygments==2.14.0
pyparsing==3.0.9
pyserial==3.5
pytest==7.2.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rospkg==1.5.0
seaborn==0.12.2
setuptools-scm==7.1.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6